pydantic_ai/requirements.txt
pydantic-ai>=0.0.32,<0.0.33
pydantic-graph>=0.0.32,<0.0.33